Add support for evcxr
This patch adds support for writing the figure to evcxr.
The functionality requires re-implementing the subprocess-spawning
machinery to apply some subtle modifications to the way in which gnuplot
output is captured. Perhaps it could be a good idea to extend the
existing code with options to suppport the new funtionality; this is
primarily just a quick hack.